pyproject.toml
[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[project]
name = "ghmcp"
version = "0.0.1"
dependencies = [
    "mcp (>=1.13.1)",
    "GitPython (>=3.1.45)"
]
requires-python = ">=3.13,<3.14"
authors = [
    { name = "Rubens Gomes", email = "rubens.s.gomes@gmail.com" }
]
maintainers = [
    { name = "Rubens Gomes", email = "rubens.s.gomes@gmail.com" }
]
description = "An AI MCP server to expose Rubens GitHub projects to AI coding assistants."
readme = "README.md"
license = "Apache-2.0"
license-files = ["LICENSE"]
keywords = [ "AI", "MCP", "coding"  ]
classifiers = [
    # How mature is this project? Common values are
    #   3 - Alpha
    #   4 - Beta
    #   5 - Production/Stable
    "Development Status :: 3 - Alpha",
    # Indicate who your project is intended for
    "Intended Audience :: General Publice",
    "Topic :: Software Development :: Resume Screener",
    # Specify the Python versions you support here.
    "Programming Language :: Python :: 3.13",
]

[project.urls]
Homepage = "https://github.com/rubensgomes/ghmcp/"
Documentation = "https://github.com/rubensgomes/ghmcp/README.md"
Repository = "https://github.com/rubensgomes/ghmcp/"

[tool.poetry]
name = "ghmcp"
packages = [
    { include = "ghmcp" }
]

[tool.poetry.group.dev.dependencies]
coverage = "^7.10.4"
pytest = "^8.4.1"
pytest-mock = "^3.14.1"
